revert(runtime): remote-agent discovery back to runtime — the runtime perceives remote agents through its own deployment file
Overturns the W3 slice of #187. The ownership call was wrong on three counts:
the agent-runtime.remote-agents config is declared in the RUNTIME deployment
file (a service declaring its outbound dependencies, like any client config);
the move never changed a runtime fact — no agent-service process exists, the
same runtime JVM kept fetching cards from the same yaml, only the module label
moved; and resolving cards for statically configured URLs is outbound-call
mechanics on the same chain as RemoteAgentInvocationService (which stayed in
runtime), not the registration/discovery capability ADR-0159 defers. The
practical regression was immediate: an agent developer running locally needed
the agent-service jar on top of agent-runtime just to wire remote tools.

Catalog, properties, refresher, and the guarded RemoteAgentConfiguration
return to agent-runtime; RemoteAgentCatalogPort is deleted; agent-service is
a skeleton again; the e2e example depends only on agent-runtime. ADR-0162 is
removed per the corpus validity-cleanup policy for overturned decisions
(adr_count 66 -> 65 across all lockstep surfaces; graph 272/441 -> 270/436).

Two fixes from the excursion are kept: the openJiuwen tool-installer bean
stays isolated behind @ConditionalOnClass (optional-dependency introspection
landmine, independent of module placement) — with the remote-agents property
guard REPEATED on the nested class, because classpath scanning registers
nested configuration classes independently of their enclosing class's
conditions; and the engine.spi dependency-whitelist ArchUnit sentinel is
unaffected.
