+++ agent-runtime/src/main/java/com/huawei/ascend/runtime/boot/RuntimeAutoConfiguration.java
@@ -1,6 +1,11 @@
 package com.huawei.ascend.runtime.boot;
 
 import com.huawei.ascend.runtime.engine.a2a.A2aAgentExecutor;
+import com.huawei.ascend.runtime.engine.a2a.A2aRemoteAgentOutboundAdapter;
+import com.huawei.ascend.runtime.engine.openjiuwen.OpenJiuwenAgentRuntimeHandler;
+import com.huawei.ascend.runtime.engine.openjiuwen.OpenJiuwenRemoteToolInstaller;
+import com.huawei.ascend.runtime.engine.service.RemoteAgentCatalog;
+import com.huawei.ascend.runtime.engine.service.RemoteAgentInvocationService;
 import com.huawei.ascend.runtime.engine.spi.AgentCardProvider;
 import com.huawei.ascend.runtime.engine.spi.AgentCards;
 import com.huawei.ascend.runtime.engine.spi.AgentRuntimeHandler;
@@ -9,6 +14,7 @@ import com.huawei.ascend.runtime.engine.spi.TrajectorySettings;
 import com.huawei.ascend.runtime.engine.spi.TrajectorySinkFactory;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.regex.Pattern;
 import org.a2aproject.sdk.server.agentexecution.AgentExecutor;
@@ -33,7 +39,9 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.SmartLifecycle;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
@@ -184,6 +192,71 @@ public class RuntimeAutoConfiguration {
     }
 
     /**
+     * Remote A2A wiring, activated only when at least one remote agent URL is
+     * configured in the runtime's own deployment file: the runtime perceives the
+     * remote agents it can call as tools through {@code agent-runtime.remote-agents},
+     * the same way any service declares its outbound dependencies. Grouping the
+     * remote beans under one guarded nested configuration keeps the
+     * {@code @ConditionalOnProperty} guard in a single place instead of
+     * repeating it on every remote bean.
+     */
+    @Configuration(proxyBeanMethods = false)
+    @ConditionalOnProperty(prefix = "agent-runtime.remote-agents.0", name = "url")
+    @EnableConfigurationProperties(RemoteAgentProperties.class)
+    public static class RemoteAgentConfiguration {
+
+        @Bean @ConditionalOnMissingBean
+        public RemoteAgentCatalog remoteAgentCatalog(RemoteAgentProperties properties) {
+            return new RemoteAgentCatalog(properties.urls());
+        }
+
+        @Bean @ConditionalOnMissingBean
+        public A2aRemoteAgentOutboundAdapter a2aRemoteAgentOutboundAdapter(RemoteAgentCatalog catalog) {
+            return new A2aRemoteAgentOutboundAdapter(catalog);
+        }
+
+        @Bean @ConditionalOnMissingBean
+        public RemoteAgentInvocationService remoteAgentInvocationService(A2aRemoteAgentOutboundAdapter outboundAdapter) {
+            return new RemoteAgentInvocationService(outboundAdapter);
+        }
+
+        @Bean @ConditionalOnMissingBean
+        public A2aAgentExecutor.RemoteSupport a2aRemoteSupport(RemoteAgentInvocationService invocationService) {
+            return new A2aAgentExecutor.RemoteSupport(invocationService);
+        }
+
+        @Bean @ConditionalOnMissingBean
+        public RemoteAgentCatalogRefresher remoteAgentCatalogRefresher(RemoteAgentCatalog catalog,
+                A2aServerExecutor executor) {
+            return new RemoteAgentCatalogRefresher(catalog, executor.executor());
+        }
+
+        /**
+         * Isolated in a nested class because the openJiuwen framework is an optional dependency:
+         * a bean method whose signature mentions openJiuwen-typed classes makes reflective
+         * introspection of the enclosing configuration throw NoClassDefFoundError in hosts
+         * without the framework. The condition is evaluated from class metadata, so this nested
+         * class is never loaded unless openJiuwen is present. The remote-agents property guard is
+         * REPEATED here because classpath scanning registers nested configuration classes
+         * independently of the enclosing class — the outer @ConditionalOnProperty does not cascade.
+         */
+        @Configuration(proxyBeanMethods = false)
+        @ConditionalOnProperty(prefix = "agent-runtime.remote-agents.0", name = "url")
+        @ConditionalOnClass(name = "com.openjiuwen.core.singleagent.BaseAgent")
+        public static class OpenJiuwenRemoteToolConfiguration {
+
+            @Bean @ConditionalOnMissingBean
+            public OpenJiuwenRemoteToolInstaller openJiuwenRemoteToolInstaller(RemoteAgentCatalog catalog,
+                    ObjectProvider<OpenJiuwenAgentRuntimeHandler> handlers) {
+                OpenJiuwenRemoteToolInstaller installer =
+                        new OpenJiuwenRemoteToolInstaller(catalog::availableToolSpecs);
+                handlers.orderedStream().forEach(handler -> handler.setRuntimeToolInstaller(installer));
+                return installer;
+            }
+        }
+    }
+
+    /**
      * Holder for the pool that runs A2A agent executions. Deliberately NOT exposed
      * as a {@code java.util.concurrent.Executor} bean: Spring Boot's
      * applicationTaskExecutor backs off when any Executor bean exists, so a broad
@@ -215,6 +288,51 @@ public class RuntimeAutoConfiguration {
                 executor.shutdownNow();
                 Thread.currentThread().interrupt();
             }
+        }
+    }
+
+    /** Polls the catalog so remote runtimes that boot later (or restart) become callable without a redeploy. */
+    public static final class RemoteAgentCatalogRefresher implements SmartLifecycle {
+        private final RemoteAgentCatalog catalog;
+        private final ExecutorService executor;
+        private final AtomicBoolean running = new AtomicBoolean();
+
+        RemoteAgentCatalogRefresher(RemoteAgentCatalog catalog, ExecutorService executor) {
+            this.catalog = catalog;
+            this.executor = executor;
+        }
+
+        @Override
+        public void start() {
+            if (running.compareAndSet(false, true)) {
+                executor.execute(this::run);
+            }
+        }
+
+        void refreshOnce() {
+            catalog.refreshPending();
+        }
+
+        private void run() {
+            while (running.get()) {
+                refreshOnce();
+                try {
+                    Thread.sleep(5_000L);
+                } catch (InterruptedException interrupted) {
+                    Thread.currentThread().interrupt();
+                    running.set(false);
+                }
+            }
+        }
+
+        @Override
+        public void stop() {
+            running.set(false);
+        }
+
+        @Override
+        public boolean isRunning() {
+            return running.get();
         }
     }
 }
