+++ agent-runtime/src/main/java/com/huawei/ascend/runtime/engine/a2a/A2aRemoteAgentOutboundAdapter.java
@@ -1,7 +1,7 @@
 package com.huawei.ascend.runtime.engine.a2a;
 
 import com.huawei.ascend.runtime.engine.service.RemoteAgentInvocationService;
-import com.huawei.ascend.runtime.engine.spi.RemoteAgentCatalogPort;
+import com.huawei.ascend.runtime.engine.service.RemoteAgentCatalog;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
@@ -39,7 +39,7 @@ public final class A2aRemoteAgentOutboundAdapter implements RemoteAgentInvocatio
     // invocations and input-required continuations instead of rebuilt on every call.
     private final Map<String, ClientTransport> transportCache = new ConcurrentHashMap<>();
 
-    public A2aRemoteAgentOutboundAdapter(RemoteAgentCatalogPort catalog) {
+    public A2aRemoteAgentOutboundAdapter(RemoteAgentCatalog catalog) {
         this(remoteAgentId -> {
             String endpoint = catalog.endpoint(remoteAgentId);
             return endpoint == null || endpoint.isBlank() ? null : new JSONRPCTransport(endpoint);
