+++ agent-runtime/src/test/java/com/huawei/ascend/runtime/boot/RuntimeRemoteAgentAutoConfigurationTest.java
@@ -1,15 +1,14 @@
-package com.huawei.ascend.service.remote;
+package com.huawei.ascend.runtime.boot;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.huawei.ascend.runtime.boot.RuntimeAutoConfiguration;
 import com.huawei.ascend.runtime.engine.AgentExecutionContext;
 import com.huawei.ascend.runtime.engine.a2a.A2aAgentExecutor;
 import com.huawei.ascend.runtime.engine.a2a.A2aRemoteAgentOutboundAdapter;
 import com.huawei.ascend.runtime.engine.openjiuwen.OpenJiuwenRemoteToolInstaller;
+import com.huawei.ascend.runtime.engine.service.RemoteAgentCatalog;
 import com.huawei.ascend.runtime.engine.service.RemoteAgentInvocationService;
 import com.huawei.ascend.runtime.engine.spi.AgentRuntimeHandler;
-import com.huawei.ascend.runtime.engine.spi.RemoteAgentCatalogPort;
 import com.huawei.ascend.runtime.engine.spi.StreamAdapter;
 import com.sun.net.httpserver.HttpServer;
 import java.io.IOException;
@@ -26,12 +25,10 @@ import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-/** The service plane owns remote-agent discovery; these tests pin its wiring and refresh behavior. */
-class RemoteAgentAutoConfigurationTest {
+class RuntimeRemoteAgentAutoConfigurationTest {
 
     private final ApplicationContextRunner contextRunner = new ApplicationContextRunner()
-            .withConfiguration(AutoConfigurations.of(
-                    RuntimeAutoConfiguration.class, RemoteAgentAutoConfiguration.class));
+            .withConfiguration(AutoConfigurations.of(RuntimeAutoConfiguration.class));
 
     @Test
     void remoteAgentUrlPropertyWiresCatalogOutboundServiceAndExecutorSupport() {
@@ -39,7 +36,7 @@ class RemoteAgentAutoConfigurationTest {
                 .withUserConfiguration(SimpleHandlerConfiguration.class)
                 .withPropertyValues("agent-runtime.remote-agents[0].url=http://localhost:18081")
                 .run(context -> {
-                    assertThat(context).hasSingleBean(RemoteAgentCatalogPort.class);
+                    assertThat(context).hasSingleBean(RemoteAgentCatalog.class);
                     assertThat(context).hasSingleBean(A2aRemoteAgentOutboundAdapter.class);
                     assertThat(context).hasSingleBean(RemoteAgentInvocationService.class);
                     assertThat(context).hasSingleBean(OpenJiuwenRemoteToolInstaller.class);
@@ -54,7 +51,8 @@ class RemoteAgentAutoConfigurationTest {
         RecordingCatalog catalog = new RecordingCatalog();
         RecordingExecutorService executor = new RecordingExecutorService();
 
-        RemoteAgentCatalogRefresher refresher = new RemoteAgentCatalogRefresher(catalog, executor);
+        RuntimeAutoConfiguration.RemoteAgentCatalogRefresher refresher =
+                new RuntimeAutoConfiguration.RemoteAgentCatalogRefresher(catalog, executor);
 
         refresher.start();
 
@@ -69,13 +67,14 @@ class RemoteAgentAutoConfigurationTest {
         HttpServer server = cardServer();
         server.start();
         try {
-            RemoteAgentAutoConfiguration configuration = new RemoteAgentAutoConfiguration();
+            RuntimeAutoConfiguration.RemoteAgentConfiguration configuration =
+                    new RuntimeAutoConfiguration.RemoteAgentConfiguration();
             RemoteAgentProperties properties =
                     new RemoteAgentProperties(List.of(
                             new RemoteAgentProperties.RemoteAgent(
                                     "http://localhost:" + server.getAddress().getPort())));
 
-            RemoteAgentCatalogPort catalog = configuration.remoteAgentCatalog(properties);
+            RemoteAgentCatalog catalog = configuration.remoteAgentCatalog(properties);
 
             assertThat(catalog.availableToolSpecs()).isEmpty();
             assertThat(catalog.pendingUrls()).containsExactly(
@@ -123,15 +122,17 @@ class RemoteAgentAutoConfigurationTest {
         HttpServer server = cardServer();
         server.start();
         try {
-            RemoteAgentAutoConfiguration configuration = new RemoteAgentAutoConfiguration();
+            RuntimeAutoConfiguration.RemoteAgentConfiguration configuration =
+                    new RuntimeAutoConfiguration.RemoteAgentConfiguration();
             RemoteAgentProperties properties =
                     new RemoteAgentProperties(List.of(
                             new RemoteAgentProperties.RemoteAgent(
                                     "http://localhost:" + server.getAddress().getPort())));
-            RemoteAgentCatalogPort catalog = configuration.remoteAgentCatalog(properties);
+            RemoteAgentCatalog catalog = configuration.remoteAgentCatalog(properties);
 
             ExecutorService executor = java.util.concurrent.Executors.newSingleThreadExecutor();
-            RemoteAgentCatalogRefresher refresher = new RemoteAgentCatalogRefresher(catalog, executor);
+            RuntimeAutoConfiguration.RemoteAgentCatalogRefresher refresher =
+                    new RuntimeAutoConfiguration.RemoteAgentCatalogRefresher(catalog, executor);
             refresher.refreshOnce();
             executor.shutdownNow();
 
@@ -143,12 +144,13 @@ class RemoteAgentAutoConfigurationTest {
 
     @Test
     void remoteAgentPropertiesExposeConfiguredUrls() {
-        RemoteAgentAutoConfiguration configuration = new RemoteAgentAutoConfiguration();
+        RuntimeAutoConfiguration.RemoteAgentConfiguration configuration =
+                new RuntimeAutoConfiguration.RemoteAgentConfiguration();
         RemoteAgentProperties properties =
                 new RemoteAgentProperties(List.of(
                         new RemoteAgentProperties.RemoteAgent("http://localhost:18081")));
 
-        RemoteAgentCatalogPort catalog = configuration.remoteAgentCatalog(properties);
+        RemoteAgentCatalog catalog = configuration.remoteAgentCatalog(properties);
 
         assertThat(catalog.availableToolSpecs()).isEmpty();
         assertThat(catalog.pendingUrls()).containsExactly("http://localhost:18081");
