+++ agent-runtime/src/test/java/com/huawei/ascend/runtime/engine/openjiuwen/OpenJiuwenRemoteToolInstallerTest.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import com.huawei.ascend.runtime.common.RuntimeIdentity;
 import com.huawei.ascend.runtime.engine.AgentExecutionContext;
-import com.huawei.ascend.runtime.engine.spi.RemoteAgentCatalogPort;
+import com.huawei.ascend.runtime.engine.service.RemoteAgentCatalog;
 import com.openjiuwen.core.foundation.llm.schema.ToolCall;
 import com.openjiuwen.core.session.AgentSessionApi;
 import com.openjiuwen.core.session.Session;
@@ -30,7 +30,7 @@ class OpenJiuwenRemoteToolInstallerTest {
     @Test
     void installsRemoteToolCardAndInterruptRailOnAgentInstance() {
         RecordingAgent agent = new RecordingAgent();
-        RemoteAgentCatalogPort.RemoteAgentToolSpec spec = toolSpec();
+        RemoteAgentCatalog.RemoteAgentToolSpec spec = toolSpec();
         OpenJiuwenRemoteToolInstaller installer = new OpenJiuwenRemoteToolInstaller(() -> List.of(spec));
 
         installer.install(agent, context());
@@ -46,7 +46,7 @@ class OpenJiuwenRemoteToolInstallerTest {
 
     @Test
     void railConvertsRemoteToolCallToInterruptContext() {
-        RemoteAgentCatalogPort.RemoteAgentToolSpec spec = toolSpec();
+        RemoteAgentCatalog.RemoteAgentToolSpec spec = toolSpec();
         AgentExecutionContext context = context();
         OpenJiuwenRemoteAgentInterruptRail rail =
                 new OpenJiuwenRemoteAgentInterruptRail(context, List.of(spec));
@@ -82,7 +82,7 @@ class OpenJiuwenRemoteToolInstallerTest {
 
     @Test
     void railTurnsRemoteResumeInputIntoSyntheticToolResult() {
-        RemoteAgentCatalogPort.RemoteAgentToolSpec spec = toolSpec();
+        RemoteAgentCatalog.RemoteAgentToolSpec spec = toolSpec();
         OpenJiuwenRemoteAgentInterruptRail rail =
                 new OpenJiuwenRemoteAgentInterruptRail(context(), List.of(spec));
         ToolCall toolCall = ToolCall.builder()
@@ -114,8 +114,8 @@ class OpenJiuwenRemoteToolInstallerTest {
         assertThat(inputs.getToolMsg().getContent()).isEqualTo("{\"ok\":true}");
     }
 
-    private static RemoteAgentCatalogPort.RemoteAgentToolSpec toolSpec() {
-        return new RemoteAgentCatalogPort.RemoteAgentToolSpec(
+    private static RemoteAgentCatalog.RemoteAgentToolSpec toolSpec() {
+        return new RemoteAgentCatalog.RemoteAgentToolSpec(
                 "remote-planner",
                 "a2a_remote_remote_planner",
                 "Remote Planner\nPlans trips",
