--- docs/adr/0162-remote-agent-discovery-to-service.yaml
@@ -1,64 +1,0 @@
-id: ADR-0162
-title: "Remote-agent discovery moves to agent-service behind a runtime-owned port (first lift of the ADR-0159 registration/discovery deferral)"
-status: accepted
-date: 2026-06-11
-level: L1
-view: development
-affects_level: L1
-affects_view: [development, logical]
-authors:
-  - chao
-extends:
-  - ADR-0159    # reserved registration/discovery for agent-service and deferred its design; this ADR lifts the first slice
-relates_to:
-  - ADR-0161    # the handler lifecycle whose serving window the catalog refresher rides on
-supersedes: []
-context: |
-  The remote-A2A-tool feature landed RemoteAgentCatalog (agent-card fetching, caching,
-  refresh, multi-instance de-duplication, tool-spec projection) plus its configuration
-  inside agent-runtime, because the committer asked for the feature to be implemented in
-  the runtime repo. ADR-0159 §Decision.7 had explicitly RESERVED registration/discovery
-  as agent-service's future enterprise capability. An architecture ownership review
-  (2026-06-11) confirmed the conflict: discovery of remote runtimes is topology
-  management — a service-plane concern — while the runtime's charter is executing one
-  agent and consuming an already-resolved view of its surroundings. Every other
-  remote-A2A piece (outbound invocation, continuation state machine, interrupt rails)
-  was confirmed runtime-owned: it is per-task execution mechanics.
-
-decision: |
-  1. agent-runtime gains one port interface, `engine.spi.RemoteAgentCatalogPort`
-     (availableToolSpecs / endpoint / refreshPending / pendingUrls + the
-     RemoteAgentToolSpec carrier record). The runtime consumes the resolved
-     remote-agent view exclusively through this port and never owns discovery.
-
-  2. The catalog implementation (card fetch via the official A2A SDK
-     A2ACardResolver, caching, refresh loop, multi-instance de-duplication,
-     tool-spec projection), the `agent-runtime.remote-agents` configuration
-     binding, and the SmartLifecycle refresher move to agent-service
-     (`com.huawei.ascend.service.remote`), wired by a service-owned
-     auto-configuration that also instantiates the runtime-side execution
-     beans (outbound adapter, invocation service, RemoteSupport, openJiuwen
-     tool installer). The dependency direction stays service -> runtime: the
-     service implements a runtime-defined port.
-
-  3. The openJiuwen tool-installer bean is isolated in a nested
-     @ConditionalOnClass configuration because the framework is an optional
-     dependency of agent-runtime (same pattern as the actuator-optional
-     health contribution).
-
-  4. Deployments that need remote-agent topology (e.g. the
-     remote-openjiuwen e2e example) now depend on agent-service. The
-     property prefix `agent-runtime.remote-agents` is kept for
-     configuration compatibility.
-
-consequences: |
-  * agent-service stops being an empty skeleton: discovery is its first real
-    capability, exactly the slice ADR-0159 reserved for it. The rest of the
-    registration/discovery design (local-agent registration, multi-agent
-    routing) remains deferred.
-  * agent-runtime's only sibling-module surface is the port it defines;
-    `engine/service` retains only the execution-side
-    RemoteAgentInvocationService.
-  * A runtime booted without agent-service simply has no
-    RemoteAgentCatalogPort bean and therefore no remote-agent tools — the
-    same behavior as an unset remote-agents property before this change.
